test: セットアップルールのテストを書きながらリファクタリング
