+++ docs/plan.md
@@ -7,9 +7,11 @@
    - ターンフロールール
    - リソース管理ルール
    - 混沌ルール
+- GameState.drawCards()が勝手に捨て札を使わないよう修正
 - PlaceCardRule.apply()で、置き場のカードの処理をそれぞれ分離する
 - 分離したものをさらにルールとして使えるようにする
 - GameStateをイミュータブルにする
+- ゲームの初期化がGameContextのコンストラクタとSetupRuleに分かれてしまっているのは統合したい
 
 ## 済み
